step-1: функция filterOrders вынесена отдельно
